Fix distance calculation in TopCode decode
The distance calculation was incorrect after removing the addition of
three nearby distance calculations. Additionally, I divided the distance
by 2 erroneously in the distance calculation.

I opted to move back to using the 3 nearby samples on the distance
calculation as it significantly improves the accuracy. I'll do some
benchmarking with flamegraph to see if there are any obvious
contributors to the recent slow down in performance.