Implement generic scan
The scan method no longer concerns itself with the implementation of the
underlying buffer. It simply uses the provided buffer and the provided
function to extract the appropriate RGB values. This saves a copy pass
from the start of the scanning process, which has led to a ~15% speed up
overall.